Updated the engine to v0.5
Fixed the Perseverance soul's web jumping, and added grazing mechanics. The problem that occured with the web jumping was that the webs weren't distanced enough to jump between. TP is being worked on now, and I can't figure out how to make multiple Grazing soul Types because the platform hates me.